@@ -20,20 +20,20 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
+        "php": "^8.1",
         "friendsofsymfony/rest-bundle": "^3.0",
-        "jms/serializer-bundle": "^3.0",
-        "symfony/framework-bundle": "^5.4",
-        "symfony/asset": "^5.4",
-        "symfony/dependency-injection": "^5.4",
-        "symfony/routing": "^5.4",
-        "symfony/security-bundle": "^5.4",
-        "symfony/yaml": "^5.4",
-        "symfony/form": "^5.4",
-        "symfony/validator": "^5.4",
-        "symfony/twig-bundle": "^5.4",
-        "symfony/twig-bridge": "^5.4",
-        "twig/twig": "^2.0|^3.0",
+        "jms/serializer-bundle": "^4.0|^5.0",
+        "symfony/framework-bundle": "^5.4|^6.4",
+        "symfony/asset": "^5.4|^6.4",
+        "symfony/dependency-injection": "^5.4|^6.4",
+        "symfony/routing": "^5.4|^6.4",
+        "symfony/security-bundle": "^5.4|^6.4",
+        "symfony/yaml": "^5.4|^6.4",
+        "symfony/form": "^5.4|^6.4",
+        "symfony/validator": "^5.4|^6.4",
+        "symfony/twig-bundle": "^5.4|^6.4",
+        "symfony/twig-bridge": "^5.4|^6.4",
+        "twig/twig": "^3.0",
         "phpoption/phpoption": "^1.1",
         "doctrine/annotations": "^1.0",
         "handcraftedinthealps/rest-routing-bundle": "^1.0"
@@ -44,14 +44,14 @@
         "symfony/assetic-bundle": "Integrates Assetic into Symfony."
     },
     "require-dev": {
-        "symfony/browser-kit": "^5.4",
-        "symfony/dom-crawler": "^5.4",
-        "symfony/css-selector": "^5.4",
+        "symfony/browser-kit": "^5.4|^6.4",
+        "symfony/dom-crawler": "^5.4|^6.4",
+        "symfony/css-selector": "^5.4|^6.4",
         "doctrine/orm": "^2.8",
         "doctrine/doctrine-bundle": "^2.4",
         "ornicar/akismet-bundle": "dev-master",
-        "symfony/expression-language": "^5.4",
-        "symfony/phpunit-bridge": "^5.4",
+        "symfony/expression-language": "^5.4|^6.4",
+        "symfony/phpunit-bridge": "^5.4|^6.4",
         "phpunit/phpunit": "^9.4",
         "sonata-project/user-bundle": "^5.10",
         "symfony/process": "^6.4"
